drfsite/requirements.txt
django
djangorestframework
markdown
django-filter
django-environ
djangorestframework-simplejwt